Update: add put method init

+++ src/main/java/com/ftalk/samsu/security/JwtAuthenticationFilter.java
@@ -31,7 +31,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 			String jwt = getJwtFromRequest(request);
 
 			if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
-				Long userId = tokenProvider.getUserIdFromJWT(jwt);
+				Integer userId = tokenProvider.getUserIdFromJWT(jwt);
 
 				UserDetails userDetails = customUserDetailsService.loadUserById(userId);
 				UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(userDetails, null,
