+++ src/main/java/com/ftalk/samsu/service/CustomUserDetailsService.java
@@ -7,6 +7,6 @@ public interface CustomUserDetailsService {
 
 	UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException;
 
-	UserDetails loadUserById(Long id);
+	UserDetails loadUserById(Integer id);
 
 }
